refactor: move getObjectFromTokens function to XliffFunctions.ts, to avoid vscode dependency where not allowed
