+++ extension/src/XliffFunctions.ts
@@ -25,7 +25,6 @@ import { Dictionary } from "./Dictionary";
 import { RefreshXlfHint, TranslationMode } from "./Enums";
 import { LanguageFunctionsSettings } from "./Settings/LanguageFunctionsSettings";
 import { RefreshResult } from "./RefreshResult";
-import { getObjectFromTokens } from "./DocumentFunctions";
 import * as ALParser from "./ALObject/ALParser";
 import { XliffIdToken } from "./ALObject/XliffIdToken";
 import { ALObject } from "./ALObject/ALElementTypes";
@@ -1420,4 +1419,23 @@ export async function createCrossLanguageXlfFromFiles(
       }
     }
   }
+}
+
+export function getObjectFromTokens(
+  alObjects: ALObject[],
+  tokens: XliffIdToken[]
+): ALObject {
+  const obj = alObjects.find(
+    (x) =>
+      x.objectType.toLowerCase() === tokens[0].type.toLowerCase() &&
+      x.objectName.toLowerCase() === tokens[0].name.toLowerCase()
+  );
+  if (!obj) {
+    throw new Error(
+      `Could not find any object matching '${XliffIdToken.getXliffIdWithNames(
+        tokens
+      )}'`
+    );
+  }
+  return obj;
 }
